각성레이아나 2번째 패턴까지 이펙트 및 뷸렛 추가

+++ Engine_SOURCE/jkAnimation.cpp
@@ -30,7 +30,7 @@ namespace jk
 		if (mbComplete)
 			return;
 
-		mTime += Time::DeltaTime();
+		mTime += static_cast<float>(Time::DeltaTime());
 
 		if (mSprites[mIndex].duration <= mTime)
 		{
